refactor(registry): split decisions PreferencesPanel under 200-line policy
